feat: Move SelectGender to own component
